Move NorthStarApplicantBaseView -> NorthStarBaseView
